image/convert/requirements.txt
libdev==0.53
Pillow==9.4.0
requests==2.28.2
